made BertLM inherit from BERT class
